+++ examples/scenario_instance/client/_api_client.py
@@ -21,7 +21,7 @@ class ScenarioInstanceClient:
     ScenarioInstanceClient
 
     Generated with:
-        pygen = 0.99.36
+        pygen = 0.99.37
         cognite-sdk = 7.63.3
         pydantic = 2.9.2
 
@@ -39,7 +39,7 @@ class ScenarioInstanceClient:
         else:
             raise ValueError(f"Expected CogniteClient or ClientConfig, got {type(config_or_client)}")
         # The client name is used for aggregated logging of Pygen Usage
-        client.config.client_name = "CognitePygen:0.99.36"
+        client.config.client_name = "CognitePygen:0.99.37"
 
         self._client = client