+++ docs/CHANGELOG.md
@@ -15,8 +15,9 @@ Changes are grouped as follows
 
 ## TBD
 ### Added
-
 - Support for deleting by `NodeId` in generated SDK.
+- You can now use `DirectRelationReference` and `tuple[str, str]` as node identifiers in the Write classes
+  of the generated SDK.
 
 ## [0.99.36] - 24-10-04
 ### Changed