+++ docs/CHANGELOG.md
@@ -13,7 +13,7 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
-## TBD
+## [0.99.37] - 24-10-15
 ### Added
 - Support for deleting by `NodeId` in generated SDK.
 - You can now use `DirectRelationReference` and `tuple[str, str]` as node identifiers in the Write classes